refactor: extract & move base64 conversion
