+++ src/shared/lib/index.ts
@@ -1,0 +1,2 @@
+export { cn } from './cn';
+export { convertFileToBase64 } from './convertToBase64';
